requirements.txt
selenium
clarifai
pyvirtualdisplay
emoji
GitPython
python-dotenv
